requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0+cu118
torchvision>=0.15.0+cu118
torchaudio>=2.0.0+cu118
transformers>=4.30.0
opencv-python>=4.8.0
mss>=9.0.1
pyttsx3>=2.90
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
regex>=2023.6.3
packaging>=23.1
typing-extensions>=4.7.1
filelock>=3.12.2
sympy>=1.12
networkx>=3.1
